# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Piggy Bank API
  description: |
    Multi-chain dust sweeper API that consolidates small token balances across EVM chains and Solana into DeFi yield positions.
    
    ## Authentication
    Most endpoints require authentication via SIWE (Sign-In with Ethereum).
    Send the signature in the `Authorization` header as `Bearer <signature>`.
    
    ## Rate Limits
    - General endpoints: 100 requests/minute
    - Quote endpoints: 50 requests/minute
    - Sweep execution: 10 requests/minute
    
    ## x402 Payments
    Some endpoints require payment via the x402 protocol. When payment is required,
    the API returns HTTP 402 with payment instructions.
  version: 1.0.0
  contact:
    name: Piggy Bank Team
    url: https://github.com/nirholas/piggy-bank
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.piggy.bank
    description: Production
  - url: http://localhost:3000
    description: Development
...
paths:
...
  /api/wallet/{address}/dust:
    get:
      tags: [Wallet]
      summary: Scan wallet for dust tokens
      description: |
        Scans a wallet across multiple chains to find small token balances ("dust").
        Can run synchronously or asynchronously for large scans.
      operationId: getWalletDust
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
            pattern: '^0x[a-fA-F0-9]{40}$'
          example: '0x742d35Cc6634C0532925a3b844Bc9e7595f8B123'
        - name: chains
          in: query
          description: Comma-separated list of chains to scan
          schema:
            type: string
          example: 'ethereum,base,arbitrum'
        - name: threshold
          in: query
          description: Maximum USD value to consider as dust
          schema:
            type: number
            default: 10
          example: 10
        - name: async
          in: query
          description: Run scan asynchronously
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Dust tokens found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DustScanResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: SIWE signature token
  schemas:
...
    DustScanResponse:
      type: object
      properties:
        wallet:
          type: string
        dust:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/DustToken'
        summary:
          type: object
          properties:
            totalTokens:
              type: integer
            totalValueUsd:
              type: number
            chains:
              type: array
              items:
                type: string
            scannedAt:
              type: integer
    DustToken:
      type: object
      properties:
        token:
          type: object
          properties:
            address:
              type: string
            symbol:
              type: string
            balance:
              type: string
            decimals:
              type: integer
        usdValue:
          type: number
        validation:
          type: object
          properties:
            canSweep:
              type: boolean
            reasons:
              type: array
              items:
                type: string
...